feat: jampage 컨벤션에 맞게 코드 분리
